semi-non-trivial changes to .openScreen, and improvements to rendering fancy 3d dots on linux
Refined how pldaps.openScreen determines width and height of screen in visual degrees
which is then used to compute pixels per degree. Used to compute width of full
display in in vis deg, then divide by screenRect width to get ppd. Mostly ok
for reasonable sized displays, but on super large projector displays the angular
extend of pixels are not uniform across the flat display. Switched it to use
the screen height instead of width for a noticeably less distorted estimate of
ppd, but still suboptimal. ...worth noting though.